requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.2
torch>=1.13.0
PyPDF2==3.0.1
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
